refactor: 마이페이지/화면 여백 관련 코드 리팩토링 및 폴더 구조 변경
