+++ app/signup/page.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Link from 'next/link';
-import { SignupForm } from './SignupForm';
+import { SignupForm } from '../../components/Form/SignupForm';
 import logo from '/public/Fittering_logo.png';
 import Image from 'next/image';
 
@@ -10,7 +10,6 @@ function SignUpPage(props: Props) {
   return (
     <div className="flex flex-col text-center mt-8 justify-center w-[29%] ml-auto mr-auto">
       <div className="mb-4 p-10 border">
-        {/* <h1 className="text-3xl font-bold">Fittering</h1> */}
         <h1 className="text-3xl font-bold">
           <Image className="w-1/3 block m-auto mt-6" src={logo} alt="Google" />
         </h1>
